fix(breadcrumbs): update breadcrumb height and stories
