+++ core/components/atoms/breadcrumbs/__stories__/moreThan4Levels.story.jsx
@@ -3,7 +3,7 @@ import { Breadcrumbs } from '@/index';
 import { action } from '@/utils/action';
 
 // CSF format story
-export const lessThan4Levels = () => {
+export const moreThan4Levels = () => {
   const list = [
     {
       label: 'Level 0',
@@ -20,6 +20,10 @@ export const lessThan4Levels = () => {
     {
       label: 'Level 3',
       link: '/level3',
+    },
+    {
+      label: 'Level 4',
+      link: '/level4',
     },
   ];
 
@@ -47,7 +51,11 @@ const customCode = `() => {
     {
       label: 'Level 3',
       link: '/level3'
-    }
+    },
+    {
+      label: 'Level 4',
+      link: '/level4'
+    },
   ];
 
   return (
@@ -61,7 +69,7 @@ const customCode = `() => {
 }`;
 
 export default {
-  title: 'Components/Breadcrumbs/Variants/Less Than 4 Levels',
+  title: 'Components/Breadcrumbs/More Than 4 Levels',
   component: Breadcrumbs,
   parameters: {
     docs: {
