+++ docs/src/pages/components/breadcrumbs/code.mdx
@@ -1,7 +1,9 @@
 #### Live Demo
-<br/>
 
-<Preview name="components-breadcrumbs-variants-less-than-4-levels--less-than-4-levels" />
+<br />
+
+<Preview name="components-breadcrumbs-less-than-4-levels--less-than-4-levels" />
 
 #### Props
-<PreviewWithPropTable name="components-breadcrumbs-variants-less-than-4-levels--less-than-4-levels" />
+
+<PreviewWithPropTable name="components-breadcrumbs-less-than-4-levels--less-than-4-levels" />
